pyproject.toml
[project]
name = "auto-apply-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "python-slugify>=8.0.0",
    "anyio>=4.0.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.2.4",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "alembic>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/auto_apply_ai"]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["src"]
